# api/manager/swagger.yaml
basePath: /
definitions:
  manager.FailureResponse:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
  manager.JobInfoResponse:
    properties:
      job:
        $ref: '#/definitions/model.Job'
        type: object
    type: object
  manager.JobListResponse:
    properties:
      jobs:
        items:
          $ref: '#/definitions/model.Job'
        type: array
    type: object
  manager.PipelineCreateResponse:
    properties:
      pipeline:
        $ref: '#/definitions/model.Pipeline'
        type: object
      status:
        type: string
    type: object
  manager.PipelineExecuteRequest:
    properties:
      name:
        type: string
    type: object
  manager.PipelineListResponse:
    properties:
      pipelines:
        items:
          $ref: '#/definitions/model.Pipeline'
        type: array
    type: object
  manager.SuccessResponse:
    properties:
      status:
        type: string
    type: object
  model.Job:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      pipeline:
        $ref: '#/definitions/model.Pipeline'
        type: object
      updatedAt:
        type: string
    type: object
  model.Pipeline:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      pipeline:
        $ref: '#/definitions/pipeline.Pipeline'
        type: object
      script:
        type: string
      updatedAt:
        type: string
    type: object
  pipeline.Command:
    properties:
      lines:
        items:
          type: string
        type: array
      name:
        type: string
    type: object
  pipeline.Expression:
    properties:
      script:
        type: string
    type: object
  pipeline.Pipeline:
    properties:
      name:
        type: string
      stages:
        items:
          $ref: '#/definitions/pipeline.Stage'
        type: array
    type: object
  pipeline.Stage:
    properties:
      environment:
        type: string
      name:
        type: string
      tasks:
        items:
          $ref: '#/definitions/pipeline.Task'
        type: array
      variables:
        items:
          $ref: '#/definitions/pipeline.Variable'
        type: array
    type: object
  pipeline.Task:
    properties:
      command:
        items:
          $ref: '#/definitions/pipeline.Command'
        type: array
      ignore_errors:
        type: boolean
      name:
        type: string
      notify:
        description: NotifyRef
        items:
          type: string
        type: array
      register:
        description: VariableRef
        type: string
      when:
        $ref: '#/definitions/pipeline.Expression'
        type: object
    type: object
  pipeline.Variable:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
info:
  contact: {}
  description: dips manager api
  license: {}
  title: dips
  version: "0.1"
paths:
  /manager/job/info/{job_id}:
    get:
      description: This method will return a single job by it's id or an error.
      operationId: job-info
      parameters:
      - description: Job ID
        in: path
        name: job_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/manager.JobInfoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/manager.FailureResponse'
      summary: find a single job by it's id
      tags:
      - jobs
  /manager/job/list:
    get:
      description: This method will return a list of all jobs
      operationId: job-list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/manager.JobListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/manager.FailureResponse'
      summary: lists all jobs
      tags:
      - jobs
  /manager/pipeline/:
    post:
      consumes:
      - text/plain
      description: This method will create the pipeline sent via the post body
      operationId: create-pipeline
      parameters:
      - description: Pipeline Script
        in: body
        name: pipeline
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/manager.PipelineCreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/manager.FailureResponse'
      summary: creates a pipeline
      tags:
      - pipelines
  /manager/pipeline/execute/{pipeline_id}:
    post:
      consumes:
      - text/plain
      description: This method will execute the pipeline with the given id
      operationId: pipeline-execute
      parameters:
      - description: Pipeline ID
        in: path
        name: pipeline_id
        required: true
        type: string
      - description: Request Body
        in: body
        name: pipeline
        required: true
        schema:
          $ref: '#/definitions/manager.PipelineExecuteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/manager.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/manager.FailureResponse'
      summary: executes a pipeline
      tags:
      - pipelines
swagger: "2.0"